+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes to the plugin.
 ### Changed
 
 - Now optionally allows `*script` commands to execute. By default the command is blocked, but you can change that through the new "Allow Unsafe Script" setting. (MR #34, courtesy Carey Williams.)
+- Updated Choicescript to latest version (3c51a79).
 
 ### Fixed